% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optim.do.simulations}
\alias{optim.do.simulations}
\title{Perform simulations as to estimate the prediction uncertainty}
\usage{
See demo
}
\arguments{
\item{fm}{is the fitted model}

\item{df}{is the dataframe with the data to which the model will be fitted}

\item{fit.func}{is the function with the fitted function}

\item{model}{is the model equation}

\item{N.sims}{is the number of simulations (repeats)}

\item{tol}{is the tolerance}

\item{trace}{controls if the function should be verbose during execution.}
}
\value{
Syntheric data based on the fitted parameters.
}
\description{
Perform simulations as to estimate the prediction uncertainty
}
\author{
Claus E. Andersen
}

