% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim-fitting-demo-function.R
\name{optim.fitting.demo}
\alias{optim.fitting.demo}
\title{Demonstration of functions in clanOptim}
\usage{
optim.fitting.demo()
}
\arguments{
\item{x}{is not used for anything}
}
\value{
Nothing expect side effects.
}
\description{
Demonstrates how optim (from the MASS library) and the functions in clanOptim
can be used for non-linear fitting.
}
\author{
Claus E. Andersen
}
