% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim-predict-function.R
\name{optim.predict}
\alias{optim.predict}
\title{Make prediction withs the fitted function.}
\usage{
fix.parameters(c(10,30,40),list(p5=50, p2=20)) # gives 10 20 30 40 50
  fix.parameters(NULL,list(p1=20))        # gives 20
  fix.parameters(c(10,20),list(p3=30))    # gives 10 20 30
  fix.parameters(c(10,20),list(p5=30))    # error - there is no 5th parameter
  fix.parameters(c(10,20),list(p3=30,x3=45))  # error - cant refer to the 3rd parameter more than once
}
\arguments{
\item{fm}{is the fitted model}

\item{df}{is the dataframe with the data to which the model will be fitted}

\item{fit.func}{is the function with the fitted function}

\item{model}{is the model equation}

\item{trace}{controls if the function should be verbose during execution.}
}
\value{
A merged parameter vector
}
\description{
Make predictions with the fitted function.
}
\author{
Claus E. Andersen
}
