% Generated by roxygen2 (4.0.1): do not edit by hand
\name{optim.start}
\alias{optim.start}
\title{Start function (see demo)}
\usage{
fix.parameters(c(10,30,40),list(p5=50, p2=20)) # gives 10 20 30 40 50
  fix.parameters(NULL,list(p1=20))        # gives 20
  fix.parameters(c(10,20),list(p3=30))    # gives 10 20 30
  fix.parameters(c(10,20),list(p5=30))    # error - there is no 5th parameter
  fix.parameters(c(10,20),list(p3=30,x3=45))  # error - cant refer to the 3rd parameter more than once
}
\arguments{
\item{df}{is the dataframe with the data to which the model will be fitted}

\item{model}{is the model equation}

\item{p.max}{is the maximum of iterations}

\item{attach.df}{controls if the dataframe will be attached or not}

\item{trace}{controls if the function should be verbose during execution.}
}
\value{
A merged parameter vector
}
\description{
Start function (see demo)
}
\author{
Claus E. Andersen
}

