% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanOptim-function.R
\name{clanOptim}
\alias{clanOptim}
\title{Version function for the clanOptim library}
\usage{
clanOptim()
}
\value{
A list of information about the version and functions within clanOptim.
}
\description{
Version function for the clanOptim library.
}
\author{
Claus E. Andersen
}
